Agregando healthz y readyz para el control y otros cambios para que corra la app
