+++ FlyLib.API/FlyLib.API.http
@@ -1,6 +1,6 @@
 @FlyLib.API_HostAddress = http://localhost:5232
 
-GET {{FlyLib.API_HostAddress}}/swagger/
+GET {{FlyLib.API_HostAddress}}/
 Accept: application/json
 
 ###
