+++ FlyLib.API/Program.cs
@@ -2,7 +2,6 @@
 using FlyLib.API.Extensions;
 using FlyLib.API.Middleware;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
-using Microsoft.OpenApi.Models;
 using Serilog;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -18,29 +17,7 @@ Log.Logger = new LoggerConfiguration()
 builder.Host.UseSerilog();
 
 builder.Services.AddControllers();
-
-//Configuraciuon Swaggerbuilder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen(c =>
-{
-    c.SwaggerDoc("v1", new OpenApiInfo
-    {
-        Title = "FlyLib API",
-        Version = "v1",
-        Description = "API para registrar viajes y fotos de FlyLib",
-        Contact = new OpenApiContact
-        {
-            Name = "Ignacio Tejerizo",
-            Email = "ignacio.tejerizo.es@gmail.com",
-            Url = new Uri("https://github.com/itejerizoes/FlyLib")
-        }
-    });
-
-    // Incluir comentarios de XML para ejemplos y docs
-    var xmlFile = $"{System.Reflection.Assembly.GetExecutingAssembly().GetName().Name}.xml";
-    var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
-    if (File.Exists(xmlPath))
-        c.IncludeXmlComments(xmlPath);
-});
+builder.Services.AddSwaggerGen();
 
 builder.Services.ConfigureOptions<ConfigureSwaggerOptions>();
 
@@ -74,4 +51,12 @@ app.UseAuthorization();
 app.UseMiddleware<GlobalExceptionMiddleware>();
 
 app.MapControllers();
+
+// Health endpoints
+app.MapHealthChecks("/healthz"); // Liveness
+app.MapHealthChecks("/readyz", new Microsoft.AspNetCore.Diagnostics.HealthChecks.HealthCheckOptions
+{
+    Predicate = check => check.Tags.Contains("ready")
+});
+
 app.Run();
